+++ CHANGELOG.md
@@ -17,6 +17,8 @@ For details about compatibility between different releases, see the **Commitment
 
 ### Changed
 
+- Improve error messages when claiming a managed gateway.
+
 ### Deprecated
 
 ### Removed